allow more shell-style redirects with /source
Previously

    /source "filename.sql" $> output.csv;

would only work if "filename.sql" was quoted, due to a quirk in how
we approach detection of valid SQL on the left-hand-side using sqlglot.

With this change, we allow any /source expression on the LHS of a
shell-style redirect operator.

The key change is a quite small addition to hybrid_redirection.py, but
requires reorganizing several methods to avoid a circular import.
