+++ mycli/client_commands.py
@@ -4,13 +4,10 @@ from collections.abc import Generator, Mapping
 import logging
 import os
 import re
-import shlex
 from typing import TYPE_CHECKING, Any, cast
 
 import click
-import sqlparse
 
-from mycli.compat import WIN
 from mycli.config import write_default_config
 from mycli.main_modes.repl import set_all_external_titles
 from mycli.packages import special
@@ -18,9 +15,12 @@ from mycli.packages.batch_utils import statements_from_filehandle
 from mycli.packages.filepaths import dir_path_exists
 from mycli.packages.interactive_utils import confirm_destructive_query
 from mycli.packages.ptoolkit.history import FileHistoryWithTimestamp
-from mycli.packages.special import main as special_main
-from mycli.packages.special.iocommands import expand_favorite_query
 from mycli.packages.special.main import ArgType, SpecialCommandAlias
+from mycli.packages.special.source import (
+    parse_source_arguments,
+    parse_source_filename,
+    source_special_command_is_safe,
+)
 from mycli.packages.sqlresult import SQLResult
 from mycli.sqlexecute import SQLExecute
 
@@ -36,136 +36,12 @@ MISSING_CONFIG_VALUE = object()
 DSN_CONFIG_VALUE = object()
 FAVORITES_CONFIG_VALUE = object()
 HIDDEN_CONFIG_SECTIONS = frozenset({'alias_dsn', 'favorite_queries'})
-INVALID_SOURCE_FILENAME = 'Source accepts exactly one filename; filenames containing spaces must be quoted.'
-SOURCE_SAFE_SPECIAL_COMMANDS = frozenset({
-    'connect',
-    'fd',
-    'fs',
-    'help',
-    'l',
-    'nowarnings',
-    'ping',
-    'prompt',
-    'redirectformat',
-    'rehash',
-    'status',
-    'tableformat',
-    'timing',
-    'use',
-    'warnings',
-    'dt',
-})
-SOURCE_SAFE_SUBCOMMANDS = {
-    'config': frozenset({'help', 'get', 'search'}),
-    'dsn': frozenset({'help', 'list', 'show', 'save', 'delete'}),
-    'favorite': frozenset({'help', 'list', 'reload', 'run', 'save', 'delete'}),
-}
 
 
 def _render_config_value(value: Any) -> str:
     if isinstance(value, list):
         return ', '.join(str(item) for item in value)
     return str(value)
-
-
-def _parse_source_arguments(arg: str) -> tuple[str, bool, bool, bool]:
-    allow_special = False
-    show_queries = False
-    page_output = False
-    filename = arg
-    while arguments := filename.split(maxsplit=1):
-        if arguments[0] == '--special':
-            allow_special = True
-        elif arguments[0] == '--show':
-            show_queries = True
-        elif arguments[0] == '--page':
-            page_output = True
-        else:
-            break
-        filename = arguments[1] if len(arguments) == 2 else ''
-    return filename, allow_special, show_queries, page_output
-
-
-def _has_unquoted_whitespace(value: str) -> bool:
-    quote: str | None = None
-    escaped = False
-    for character in value:
-        if escaped:
-            if quote is None and character.isspace():
-                return True
-            escaped = False
-            continue
-        if not WIN and character == '\\' and quote != "'":
-            escaped = True
-            continue
-        if character in ("'", '"'):
-            if quote is None:
-                quote = character
-            elif quote == character:
-                quote = None
-        elif quote is None and character.isspace():
-            return True
-    return False
-
-
-def _parse_source_filename(filename: str) -> str:
-    if not filename:
-        return ''
-    if _has_unquoted_whitespace(filename):
-        raise ValueError(INVALID_SOURCE_FILENAME)
-    try:
-        arguments = shlex.split(filename, posix=not WIN)
-    except ValueError as error:
-        raise ValueError(f'Invalid source filename: {error}.') from None
-    if len(arguments) != 1:
-        raise ValueError(INVALID_SOURCE_FILENAME)
-    parsed_filename = arguments[0]
-    if WIN and len(parsed_filename) >= 2 and parsed_filename[0] == parsed_filename[-1] and parsed_filename[0] in ("'", '"'):
-        parsed_filename = parsed_filename[1:-1]
-    return parsed_filename
-
-
-def _registered_special_command(query: str) -> tuple[str, str] | None:
-    command, _verbosity, arg = special.parse_special_command(query)
-    registered = special_main.COMMANDS.get(command)
-    if registered is None:
-        registered = special_main.COMMANDS.get(command.lower())
-    if registered is None:
-        return None
-    return registered.command.removeprefix('\\').removeprefix('/').lower(), arg
-
-
-def _favorite_source_command_is_safe(arg: str) -> bool:
-    query, _error = expand_favorite_query(arg)
-    if query is None:
-        return True
-    return not any(special.is_special_command(statement.rstrip(';')) for statement in sqlparse.split(query))
-
-
-def _source_special_command_is_safe(query: str) -> bool:
-    parsed = _registered_special_command(query)
-    if parsed is None:
-        return False
-
-    command, arg = parsed
-    if command == 'f':
-        return not arg or _favorite_source_command_is_safe(arg)
-    if command in ('fd', 'fs'):
-        return True
-    if command in SOURCE_SAFE_SPECIAL_COMMANDS:
-        return True
-
-    subcommands = SOURCE_SAFE_SUBCOMMANDS.get(command)
-    if subcommands is None:
-        return False
-    arguments = arg.split(maxsplit=1)
-    subcommand = arguments[0].lower() if arguments else 'help'
-    if subcommand not in subcommands:
-        return False
-    if command == 'favorite' and subcommand == 'run':
-        run_arg = arguments[1] if len(arguments) == 2 else ''
-        return not run_arg or _favorite_source_command_is_safe(run_arg)
-    return True
 
 
 def _iter_config_values(
@@ -412,11 +288,11 @@ class ClientCommandsMixin:
         yield SQLResult(status=msg)
 
     def execute_from_file(self, arg: str, **_) -> Generator[SQLResult, None, None]:
-        filename, allow_special, show_queries, page_output = _parse_source_arguments(arg)
+        filename, allow_special, show_queries, page_output = parse_source_arguments(arg)
         if page_output:
             yield SQLResult(command={'name': 'source_page'})
         try:
-            filename = _parse_source_filename(filename)
+            filename = parse_source_filename(filename)
         except ValueError as error:
             yield SQLResult(status=str(error), is_error=True)
             return
@@ -450,7 +326,7 @@ class ClientCommandsMixin:
                             is_error=True,
                         )
                         return
-                    if not _source_special_command_is_safe(special_query):
+                    if not source_special_command_is_safe(special_query):
                         command, _verbosity, _arg = special.parse_special_command(special_query)
                         yield SQLResult(
                             status=f'Special command is never permitted in source files: {command}.',
